new public api routine: JS9.AddDivs() adds displays and/or plugins to the page
This takes the place of using two internal routines to init a JS9 div:

  new JS9.Display(id)
  JS9.InstantiatePlugins()

You can now create a JS9 display div on the fly and instantiate it, along
with its plugins.  If no id(s) is specified, the routine just performs
JS9.InstantiatePlugins().

Along the way, JS9.InstantiatePlugins was renamed to JS9.instantiatePlugins,
to make it more obvious that this is an internal routine. Same for JS9.Init(),
renamed to JS9.init().