add support for GetImageData() returning data as a base64-encoded string
Oddly, this seems to be the fastest way of transferring data to an
external process such as Python, i.e. faster than allowing socket.io (v1.2)
to processes binary data. This may change as socket.io evolves, so the
"array" method may be preferred in the future.

Also, un-named data files are now called '[anonymous]'.