change $filename macro so that it is prepended by $JS9_DIR when relative
The $filename macro, used as part of analysis command lines, points to
the FITS file pathname. Often this is path is relative to the JS9
install directory. In order to support analysis in individual working
directories, we can use a relative path. So we changed $filename so
that it prepends the string $JS9_DIR to the path.  This environment
variable is then set by the back-end server so that it is defined when
the analysis wrapper script is run.