Add links to commands
This adds an id to each command. The id is taken to be JS9.<command> since
this is unique (and doesn't clash with more-common tokens like load, which
is already used as an id). The ids are attached to the "summary" paragraph.

Most text that refers to a function has been turned into a link, but some
may be missing (and I have tried not to lnk to the function currently
being discussed).

At least one link is fixed (in the text about shapes there was a link to
AddRegions which has been changed to AddShapes).

A number of open/close p tags are added (many more could be).