implement statusbar dictionary inside js9.js
I don't want to make useStatusbarDictionary a global option because the expandMacro function is used in multiple places and I want the expansion to work only when expandMacro is called from a particular point in the statusbar plugin. So instead of a global option I made a global variable which can be set and uset as needed by the plugin code.